pyproject.toml
[project]
name = "dnd-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "google-adk>=1.27.2",
    "litellm>=1.82.4",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
]
